refactor(qe): rename wikidata query executer
