+++ pom.xml
@@ -61,7 +61,7 @@
                             <module>qanary_component-QC-EAT-classifier</module>
                             <module>qanary_component-ned-opentapioca</module>
                             <module>qanary_component-qb-birthdata-wikidata</module>
-                            <module>qanary_component-qe-birthplace</module>
+                            <module>qanary_component-qe-wikidata</module>
 			    <module>qanary_component-QBE-QAnswer</module>
 			</modules>
 		</profile>
