+++ qanary_component-qe-wikidata/Dockerfile
