+++ qanary_component-qe-wikidata/pom.xml
@@ -4,7 +4,7 @@
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 	<groupId>eu.wdaqua.qanary.component</groupId>
-	<artifactId>qanary_component-qe-birthdata-wikidata</artifactId>
+	<artifactId>qanary_component-qe-wikidata</artifactId>
 	<version>1.0.0</version>
 
 	<parent>
@@ -17,7 +17,7 @@
 		<qanary.version>[2.0.0,3.0.0)</qanary.version>
 		<spring-boot-admin.version>2.1.6</spring-boot-admin.version>
 		<docker.image.prefix>qanary</docker.image.prefix>
-		<docker.image.name>qe-birthdata-wikidata</docker.image.name>
+		<docker.image.name>qe-wikidata-executer</docker.image.name>
 		<dockerfile-maven-version>1.4.13</dockerfile-maven-version>
 	</properties>
 
@@ -93,7 +93,7 @@
 				<groupId>org.springframework.boot</groupId>
 				<artifactId>spring-boot-maven-plugin</artifactId>
 				<configuration>
-					<mainClass>eu.wdaqua.queryexecutor.Application</mainClass>
+					<mainClass>eu.wdaqua.queryexecuter.Application</mainClass>
 				</configuration>
 				<executions>
 					<execution>
