+++ qanary_component-qe-wikidata/src/main/resources/banner.txt
