+++ qanary_component-qe-wikidata/src/main/resources/config/application.properties
@@ -1,7 +1,7 @@
 
 # Update the port number
 server.port=5557
-spring.application.name=WikidataQueryExecutor
+spring.application.name=WikidataQueryExecuter
 spring.application.description=${spring.application.name} is a Qanary component
 
 # Update the URL of the Qanary pipeline
